Added RBAC based access control.

+++ app/controllers/BusinessPartners.java
@@ -5,12 +5,15 @@ import java.util.List;
 import models.BusinessPartner;
 import models.BusinessYear;
 import models.Company;
+import models.Permission;
+import models.User;
 import play.Logger;
 import play.mvc.Controller;
 
 public class BusinessPartners extends Controller {
 	
 	public static void show(String mode) {
+		authorize("viewBusinessPartners");
 		List<BusinessPartner> partners = BusinessPartner.findAll();
 		List<Company> companies = Company.findAll();
 		if (mode == null || mode.equals(""))
@@ -19,6 +22,7 @@ public class BusinessPartners extends Controller {
 	}
 
 	public static void create(BusinessPartner businesspartner) {
+		authorize("createBusinessPartner");
 		validation.required("name", businesspartner.name);
 		validation.required("kind", businesspartner.kind);
 		validation.required("account", businesspartner.account);
@@ -37,6 +41,7 @@ public class BusinessPartners extends Controller {
 	}
 
 	public static void edit(BusinessPartner businesspartner) {
+		authorize("editBusinessPartner");
 		validation.required("name", businesspartner.name);
 		validation.required("kind", businesspartner.kind);
 		validation.required("account", businesspartner.account);
@@ -67,6 +72,7 @@ public class BusinessPartners extends Controller {
 	}
 
 	public static void delete(Long id) {
+		authorize("deleteBusinessPartner");
 		if (id != null) {
 			BusinessPartner partners = BusinessPartner.findById(id);
 			partners.delete();
@@ -75,6 +81,26 @@ public class BusinessPartners extends Controller {
 			Logger.info(code + " : user = "+user + " id = "+id);
 		}
 		show("edit");
+	}
+	
+	private static void authorize(String operationName){
+		String username = Security.connected();
+		List<User> users = User.find("byUsername", username).fetch();
+		if(users.isEmpty()) {
+			render("errors/401.html");
+		} 
+		User user = users.get(0);
+		boolean check = false;
+		List<Permission> permissions = user.role.permissions;
+		for(Permission p : permissions){
+			if(p.name.equals(operationName)){
+				check = true;
+				break;
+			}
+		}
+		if(!check) {
+			render("errors/401.html");
+		}
 	}
 
 }
