+++ app/controllers/PricelistItems.java
@@ -3,14 +3,17 @@ package controllers;
 import java.util.List;
 
 import models.Item;
+import models.Permission;
 import models.Pricelist;
 import models.PricelistItem;
+import models.User;
 import play.Logger;
 import play.mvc.Controller;
 
 public class PricelistItems extends Controller{
 	
 	public static void show(String mode) {
+		authorize("viewPriceListItems");
 		List<PricelistItem> pricelistitems = PricelistItem.findAll();
 		List<Item> items = Item.findAll();		
 		List<Pricelist> pricelists = Pricelist.findAll();
@@ -20,6 +23,7 @@ public class PricelistItems extends Controller{
 	}
 	
 	public static void create(PricelistItem pricelistitem) {	
+		authorize("createPriceListItem");
 		validation.required("price", pricelistitem.price);
 		validation.min("price", pricelistitem.price, 0.01);		
 		if (validation.hasErrors()) {
@@ -37,6 +41,7 @@ public class PricelistItems extends Controller{
 	}
 
 	public static void edit(PricelistItem pricelistitem) {
+		authorize("editPriceListItem");
 		validation.required("price", pricelistitem.price);
 		validation.min("price", pricelistitem.price, 0.01);		
 		if (validation.hasErrors()) {
@@ -60,6 +65,7 @@ public class PricelistItems extends Controller{
 	}
 
 	public static void delete(Long id) {
+		authorize("deletePriceListItem");
 		if (id != null) {
 			PricelistItem pricelistitem = PricelistItem.findById(id);
 			pricelistitem.delete();
@@ -68,6 +74,26 @@ public class PricelistItems extends Controller{
 			Logger.info(code + " : user = "+user + " id = "+pricelistitem.id);
 		}
 		show("edit");
+	}
+	
+	private static void authorize(String operationName){
+		String username = Security.connected();
+		List<User> users = User.find("byUsername", username).fetch();
+		if(users.isEmpty()) {
+			render("errors/401.html");
+		} 
+		User user = users.get(0);
+		boolean check = false;
+		List<Permission> permissions = user.role.permissions;
+		for(Permission p : permissions){
+			if(p.name.equals(operationName)){
+				check = true;
+				break;
+			}
+		}
+		if(!check) {
+			render("errors/401.html");
+		}
 	}
 
 }
